Revert "feat: revert isomorphic wasi  (#1582)"
This reverts commit 230a78888abb0da56ef7cbf4d3b3703eaed607bf.
