+++ apps/wing-playground/package.json
@@ -10,13 +10,12 @@
   },
   "author": "Monada",
   "dependencies": {
-    "@wasmer/wasi": "^1.2.2",
+    "@cowasm/memfs": "^3.5.1",
+    "winglang": "file:../../apps/wing",
     "@winglang/sdk": "file:../../libs/wingsdk"
   },
   "devDependencies": {
-    "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
-    "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
-    "rollup-plugin-polyfill-node": "^0.12.0",
-    "vite": "^4.1.1"
+    "vite": "^4.1.1",
+    "vite-plugin-node-polyfills": "^0.7.0"
   }
 }
