+++ apps/wing/bin/wing
@@ -1,31 +1,3 @@
 #!/usr/bin/env node
 
-// Only certain builds and versions of Node support WASI.
-// We first assume this script is loaded in the correct runtime environment, and
-// if not, we try to load it in a WASI-compatible runtime environment by running
-// the same script (self) over a new Node process with the correct flags passed.
-// package.json warns if user npm installs this for a non-compatible runtime.
-
-try {
-  require.resolve("wasi");
-} catch (err) {
-  if (err.code === "MODULE_NOT_FOUND") {
-    const { spawnSync } = require("child_process");
-    const { env, execPath, execArgv, argv } = process;
-    env.NODE_OPTIONS = [
-      ...new Set((env.NODE_OPTIONS ?? "").split(" "))
-        .add("--experimental-wasi-unstable-preview1")
-        .add("--no-warnings"),
-    ].join(" ");
-
-    const { status } = spawnSync(execPath, execArgv.concat(argv.slice(1)), {
-      env,
-      stdio: "inherit",
-    });
-    process.exit(status ?? 1);
-  } else {
-    console.error('Unable to load "wasi" module', err);
-    process.exit(1);
-  }
-}
-require("../dist/index.js");
+require("../dist/cli.js");
